requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy~=1.19.5
pandas~=1.1.5
mlflow~=1.22.0


pyyaml~=6.0
sklearn~=0.0
scikit-learn~=0.24.2
setuptools~=58.0.4